requirements.txt
Django==1.8.3
Pillow==2.9.0
PyJWT==1.4.0
django-appconf==1.0.1
django-geoposition==0.2.2
oauthlib==1.0.1
psycopg2==2.6.1
python-openid==2.2.5
python-social-auth==0.2.12
qrcode==5.1
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
